convert DSN uploader to a function
