+++ auto-kol/agent/src/services/agents/nodes.ts
@@ -5,7 +5,7 @@ import { ChromaService } from '../vectorstore/chroma.js';
 import * as db from '../database/index.js';
 import { flagBackSkippedTweet, getAllSkippedTweetsToRecheck } from '../../database/index.js';
 import { tweetSearchSchema } from '../../schemas/workflow.js';
-
+import { uploadToDsn } from '../../utils/dsn.js';
 export const createNodes = async (config: WorkflowConfig) => {
 
     ///////////TIMELINE///////////
@@ -171,6 +171,15 @@ export const createNodes = async (config: WorkflowConfig) => {
                     .invoke({ tweet: tweet.text });
 
                 logger.info('LLM decision:', { decision });
+
+                // TODO: Add the previous cid to the dsn upload
+                await uploadToDsn({ 
+                    data: {
+                        decision,
+                        tweet
+                    }, 
+                    previousCid: '' 
+                });
 
                 if (!decision.shouldEngage) {
                     logger.info('Queueing skipped tweet to review queue:', { tweetId: tweet.id });
