+++ auto-kol/agent/src/tools/index.ts
@@ -3,7 +3,6 @@ import { createTweetSearchTool } from './tools/tweetSearchTool.js';
 import { createQueueResponseTool } from './tools/queueResponseTool.js';
 import { createQueueSkippedTool } from './tools/queueSkippedTool.js';
 import { createSearchSimilarTweetsTool } from './tools/searchSimilarTweetsTool.js';
-import { createDsnUploadTool } from './tools/dsnUploadTool.js';
 
 export const createTools = (scraper: any) => {
 
@@ -17,22 +16,18 @@ export const createTools = (scraper: any) => {
 
     const searchSimilarTweetsTool = createSearchSimilarTweetsTool();
 
-    const dsnUploadTool = createDsnUploadTool();
-
     return {
         tweetSearchTool,
         queueResponseTool,
         queueSkippedTool,
         searchSimilarTweetsTool,
         fetchTimelineTool,
-        dsnUploadTool,
         tools: [
             tweetSearchTool,
             queueResponseTool,
             queueSkippedTool,
             searchSimilarTweetsTool,
             fetchTimelineTool,
-            dsnUploadTool
         ]
     };
 };
